chore : placeId Long -> double로 변경

+++ src/main/java/com/book_everywhere/domain/pin/Pin.java
@@ -31,7 +31,7 @@ public class Pin {
     private List<Tagged> tags;
 
     @Column(nullable = false)
-    private Long placeId;
+    private double placeId;
 
     //경도 x
     @Column(nullable = false)
